+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added *Compare Numbers* node.
 - Added *Copy Object Modifiers* node.
 - Added *Export Headers* option to the setup script.
+- Added *Indices* output to *Get Random List Elements* node.
 - Added *Marching Squares* node.
 
 ### Fixed
@@ -18,6 +19,8 @@
 - Fixed nextBoolean method for XoShiRo256StarStar generators.
 - Fixed OpenGL fragment shaders on Core contexts.
 - Fixed triangulateMesh method of Mesh.
+- Fixed freeze in Is Inside Volume node.
+- Fixed Grease Pencil Nodes for API changes.
 
 ### Changed